refactor: albums
