+++ src/albums/entity/albums.entity.ts
@@ -1,9 +1,16 @@
-import { Column, Entity } from 'typeorm'
+import {
+  Column,
+  Entity,
+  JoinColumn,
+  ManyToOne,
+  PrimaryGeneratedColumn,
+} from 'typeorm'
 import { IsInt, IsUUID } from 'class-validator'
+import { ArtistEntity } from '../../artists/entity/artists.entity'
 
 @Entity('album')
 export class AlbumEntity {
-  @Column({ primary: true })
+  @PrimaryGeneratedColumn('uuid')
   id: string
 
   @Column()
@@ -15,7 +22,17 @@ export class AlbumEntity {
 
   @IsUUID()
   @Column({ nullable: true })
-  artistId: string
+  artistId: string | null
+
+  @ManyToOne(() => ArtistEntity, (artist) => artist.id, {
+    onDelete: 'SET NULL',
+    nullable: true,
+  })
+  @JoinColumn({
+    name: 'artistId',
+    referencedColumnName: 'id',
+  })
+  artist: string | null
 
   constructor(partial: Partial<AlbumEntity>) {
     Object.assign(this, partial)
