--- src/tracks/dto/update-track.dto.ts
@@ -1,13 +1,0 @@
-import { IsInt, IsNotEmpty, IsString } from 'class-validator'
-
-export class UpdateTrackDto {
-  @IsString()
-  @IsNotEmpty()
-  readonly name: string
-
-  @IsInt()
-  @IsNotEmpty()
-  readonly duration: number
-  readonly artistId: string | null // refers to Artist
-  readonly albumId: string | null // refers to Album
-}
